[ДОБАВЛЕНО] отображение постов в ленте `feed`
