+++ app/views/feed.scala.html
@@ -7,6 +7,6 @@
 
 <ul>
 @for(post <- feed) {
-  <li>post</li>
+  <li>@post</li>
 }
 </ul>
